move fetch to app.js
