+++ pages/_app.js
@@ -1,9 +1,10 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import GlobalStyles from "../components/GlobalStyles";
 import Layout from "../components/Layout/Layout";
 
 function MyApp({ Component, pageProps }) {
   const [isFavourite, setIsFavourite] = useState([]);
+  const [entries, setEntries] = useState([]);
 
   function handleFav(id, event) {
     event.preventDefault();
@@ -14,11 +15,25 @@ function MyApp({ Component, pageProps }) {
       : setIsFavourite((exercise) => [...exercise, id]);
   }
 
+  async function getExercises() {
+    const response = await fetch("/api/exercises");
+    const exercisesList = await response.json();
+    setEntries(exercisesList);
+  }
+
+  useEffect(() => {
+    getExercises();
+  }, []);
   return (
     <>
       <GlobalStyles />
       <Layout>
-        <Component {...pageProps} onFav={handleFav} isFavourite={isFavourite} />
+        <Component
+          {...pageProps}
+          onFav={handleFav}
+          isFavourite={isFavourite}
+          entries={entries}
+        />
       </Layout>
     </>
   );
