+++ pages/favourites.js
@@ -2,7 +2,7 @@ import Head from "next/head";
 import { Fragment } from "react";
 import Exercises from "../components/Main/Exercises/Exercises";
 
-const Favourites = ({ onFav, isFavourite }) => {
+const Favourites = ({ onFav, isFavourite, entries }) => {
   return (
     <Fragment>
       <Head>
@@ -16,6 +16,7 @@ const Favourites = ({ onFav, isFavourite }) => {
         onFav={onFav}
         isFavourite={isFavourite}
         isFavouriteExercise={true}
+        entries={entries}
       />
     </Fragment>
   );
